Updated package.json and other relevant files to use 'concurrently'.
