+++ .env
@@ -1,3 +1,4 @@
 MONGO_URI = mongodb+srv://username:pass1word@cluster0.hjyyb.mongodb.net/StudySphere?retryWrites=true&w=majority&appName=Cluster0
 NODE_ENV = development
 PORT = 5678
+BROWSER = none
