+++ package.json
@@ -5,6 +5,7 @@
   "main": "backend/server.js",
   "dependencies": {
     "axios": "^1.7.7",
+    "concurrently": "^9.1.0",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "express-async-handler": "^1.2.0",
@@ -28,7 +29,9 @@
   },
   "scripts": {
     "test": "jest --coverage",
-    "start": "node server.js",
+    "start": "concurrently npm:frontend npm:backend",
+    "frontend": "node react-server.js",
+    "backend": "node server.js",
     "server": "nodemon server.js"
   },
   "eslintConfig": {
